renamed script to exclude error in imports; added option to add params to sklearn score
